Fix: Delete all images before placing masks
Mask layer updating now deletes all previous mask layers. This ensures
that these layers are invalidated and empty when drawing new masks. To
do this, (1) for the functional image, look for the Image primitive with
direct alpha mapping, and (2) for the anatomy image, just delete
everything and replot the anatomy image.

This is a hacky fix but it works with every use case and it is cleaner
than placing a listbox on the analyzer manipulation panel.